solved empty df bug in error_sing_traj
There was a bug where the preds_cp and trues_cp where sliced when coming from a missing trajectory. In that case, we assign just one value, which was then cropped out, resulting in an empty CPs array, which then gives rise to perfect scoring. Now slicing only occurs when for not missing trajectories.